requirements.txt
appdirs==1.4.3
Django==1.10.6
packaging==16.8
pyparsing==2.2.0
six==1.10.0
psycopg2==2.7
djangorestframework=3.5.4
